+++ internal/provider/data_source_catalystcenter_lan_automation.go
@@ -72,7 +72,7 @@ func (d *LANAutomationDataSource) Schema(ctx context.Context, req datasource.Sch
 				Computed:            true,
 			},
 			"peer_device_management_ip_address": schema.StringAttribute{
-				MarkdownDescription: "The list of interfaces on primary seed via which the discovered devices are connected.",
+				MarkdownDescription: "Secondary seed management IP address.",
 				Computed:            true,
 			},
 			"primary_device_interface_names": schema.ListAttribute{