+++ internal/provider/resource_catalystcenter_lan_automation_test.go
@@ -43,7 +43,6 @@ func TestAccCcLANAutomation(t *testing.T) {
 	checks = append(checks, resource.TestCheckResourceAttr("catalystcenter_lan_automation.test", "ip_pools.0.ip_pool_role", "MAIN_POOL"))
 	checks = append(checks, resource.TestCheckResourceAttr("catalystcenter_lan_automation.test", "multicast_enabled", "true"))
 	checks = append(checks, resource.TestCheckResourceAttr("catalystcenter_lan_automation.test", "host_name_prefix", "TEST"))
-	checks = append(checks, resource.TestCheckResourceAttr("catalystcenter_lan_automation.test", "host_name_file_id", "1234"))
 	checks = append(checks, resource.TestCheckResourceAttr("catalystcenter_lan_automation.test", "isis_domain_password", "cisco123"))
 	checks = append(checks, resource.TestCheckResourceAttr("catalystcenter_lan_automation.test", "redistribute_isis_to_bgp", "true"))
 
@@ -99,7 +98,6 @@ func testAccCcLANAutomationConfig_all() string {
 	config += `	}]` + "\n"
 	config += `	multicast_enabled = true` + "\n"
 	config += `	host_name_prefix = "TEST"` + "\n"
-	config += `	host_name_file_id = "1234"` + "\n"
 	config += `	isis_domain_password = "cisco123"` + "\n"
 	config += `	redistribute_isis_to_bgp = true` + "\n"
 	config += `}` + "\n"